Refactor source
Use storybook for the example
Update the README
