+++ .storybook/config.js
@@ -1,0 +1,2 @@
+import * as storybook from '@kadira/storybook';
+storybook.configure(() => require('./stories'), module);
