+++ package.json
@@ -4,8 +4,9 @@
   "description": "Story Links addon for storybook",
   "main": "dist/index.js",
   "scripts": {
+    "deploy-storybook": "storybook-to-ghpages",
     "prepublish": "node .scripts/npm-prepublish.js",
-    "start": "reactpack -d --no-lint src/example/index.js",
+    "storybook": "start-storybook -p 9001",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "repository": {
@@ -21,6 +22,7 @@
   },
   "homepage": "https://github.com/kadirahq/storybook-addon-links#readme",
   "devDependencies": {
+    "@kadira/storybook": "^2.2.0",
     "@kadira/storybook-addons": "^1.3.0",
     "@kadira/storybook-ui": "^3.0.0",
     "babel-cli": "^6.11.4",
@@ -29,10 +31,11 @@
     "babel-preset-stage-0": "^6.5.0",
     "react": "^15.3.0",
     "react-dom": "^15.3.0",
-    "reactpack": "^0.7.3",
     "shelljs": "^0.7.3"
   },
   "peerDependencies": {
-    "@kadira/storybook-addons": "^1.3.0"
+    "@kadira/storybook-addons": "^1.3.0",
+    "react": "^0.14.7 || ^15.0.0",
+    "react-dom": "^0.14.7 || ^15.0.0"
   }
 }
