--- src/example/addons.js
@@ -1,3 +1,0 @@
-// register the action logger addon for storybook
-// import '@kadira/storybook-addon-links/register'
-require('../').register();
