--- src/example/index.ejs
@@ -1,14 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <meta http-equiv="Content-type" content="text/html; charset=utf-8"/>
-    <title><%= htmlWebpackPlugin.options.title %></title>
-    <style>body {margin: 0}</style>
-    <% if (htmlWebpackPlugin.options.dev) { %>
-    <script src="http://localhost:<%= htmlWebpackPlugin.options.port %>/webpack-dev-server.js"></script>
-    <% } %>
-  </head>
-  <body>
-    <div id="react-app"></div>
-  </body>
-</html>
